adjustments migrations

+++ src/main/resources/db/migration/V1_11__create_table_customer.sql
@@ -5,6 +5,6 @@ create table if not exists customer(
     social_security_number_customer varchar(11) unique not null,
     type_customer varchar(4) not null,
     status_data_base varchar(20) default 'ACTIVE',
-    rf_bank bigint(20) not null,
+    rf_bank int(3) not null,
     foreign key(rf_bank) references bank(id_bank)
 ) engine innodb default charset=utf8;
